# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cairn API
  version: 0.1.0
  description: Versioned API shared by Cairn desktop, PWA, and future mobile clients.
  license:
    name: GPL-3.0-only
servers:
  - url: http://127.0.0.1:7381
paths:
...
  /api/v1/devices/pair:
    post:
      operationId: pairDevice
      tags: [devices]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, platform]
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 12
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                platform:
                  type: string
                  enum: [web, ipad, windows, macos, ios, android]
      responses:
        "201":
          description: Device paired. Token is returned only once.
          content:
            application/json:
              schema:
                type: object
                required: [device, token]
                properties:
                  device:
                    $ref: "#/components/schemas/Device"
                  token:
                    type: string
        "400":
          $ref: "#/components/responses/Problem"
...
  /api/v1/devices/{deviceId}:
    delete:
      operationId: revokeDevice
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "204":
          description: Device token revoked.
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/ai/profiles:
    get:
      operationId: listAIProfiles
      tags: [ai]
      responses:
        "200":
          description: Configured AI profiles without API keys.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AIProfile"
    post:
      operationId: createAIProfile
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAIProfileRequest"
      responses:
        "201":
          description: Encrypted AI profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIProfile"
        "428":
          $ref: "#/components/responses/Problem"
  /api/v1/ai/profiles/{profileId}:
    parameters:
      - $ref: "#/components/parameters/AIProfileId"
    patch:
      operationId: updateAIProfile
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAIProfileRequest"
      responses:
        "200":
          description: AI profile updated without exposing its API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIProfile"
        "404":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteAIProfile
      tags: [ai]
      responses:
        "204":
          description: AI profile deleted. Existing results and chats remain without the profile link.
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/entries/{entryId}/ai-results:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    get:
      operationId: listAIResults
      tags: [entries, ai]
      responses:
        "200":
          description: Cached AI results for the article.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AIResult"
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/entries/{entryId}/ai/{operation}:
    parameters:
      - $ref: "#/components/parameters/EntryId"
      - name: operation
        in: path
        required: true
        schema:
          type: string
          enum: [summary, translation, key-points]
    post:
      operationId: runAIOperation
      tags: [entries, ai, jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile_id:
                  type: string
                  format: uuid
                language:
                  type: string
                  maxLength: 40
                  default: auto
      responses:
        "200":
          description: Matching cached result returned without a provider request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIStartResponse"
        "202":
          description: AI operation job accepted or an equivalent pending job returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIStartResponse"
        "428":
          $ref: "#/components/responses/Problem"
  /api/v1/entries/{entryId}/ai-chat:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    post:
      operationId: startAIChat
      tags: [entries, ai, jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                profile_id:
                  type: string
                  format: uuid
                session_id:
                  type: string
                  format: uuid
                message:
                  type: string
                  minLength: 1
                  maxLength: 4000
      responses:
        "202":
          description: Article chat job accepted.
          content:
            application/json:
              schema:
                type: object
                required: [job, session]
                properties:
                  job:
                    $ref: "#/components/schemas/Job"
                  session:
                    $ref: "#/components/schemas/AIChatSession"
        "428":
          $ref: "#/components/responses/Problem"
  /api/v1/ai/chats/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getAIChat
      tags: [ai]
      responses:
        "200":
          description: Persisted article chat session and messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIChatSession"
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: CairnDeviceToken
  parameters:
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AIProfileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Job:
      type: object
      required: [id, kind, state, progress_current, progress_total, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        progress_current:
          type: integer
          minimum: 0
        progress_total:
          type: integer
          minimum: 0
        error_code:
          type: [string, "null"]
        error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Device:
      type: object
      required: [id, name, platform, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
        last_seen_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: [string, "null"]
          format: date-time
...
    AISettings:
      type: object
      properties:
        temperature:
          type: number
          minimum: 0
          maximum: 2
    CreateAIProfileRequest:
      type: object
      required: [provider, endpoint, model]
      properties:
        provider:
          type: string
          enum: [openai_compatible, ollama]
        name:
          type: string
          maxLength: 120
        endpoint:
          type: string
          format: uri
        model:
          type: string
          maxLength: 200
        api_key:
          type: string
          format: password
          writeOnly: true
        settings:
          $ref: "#/components/schemas/AISettings"
        enabled:
          type: boolean
          default: true
        allow_private_network:
          type: boolean
          default: false
        remote_content_approved:
          type: boolean
          default: false
        is_default:
          type: boolean
          default: false
    UpdateAIProfileRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          maxLength: 120
        endpoint:
          type: string
          format: uri
        model:
          type: string
          maxLength: 200
        api_key:
          type: string
          format: password
          writeOnly: true
        settings:
          $ref: "#/components/schemas/AISettings"
        enabled:
          type: boolean
        allow_private_network:
          type: boolean
        remote_content_approved:
          type: boolean
        is_default:
          type: boolean
    AIProfile:
      type: object
      required: [id, provider, name, endpoint, model, enabled, allow_private_network, remote_content_approved, is_default, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [openai_compatible, ollama]
        name:
          type: string
        endpoint:
          type: string
          format: uri
        model:
          type: string
        enabled:
          type: boolean
        allow_private_network:
          type: boolean
        remote_content_approved:
          type: boolean
        is_default:
          type: boolean
        last_used_at:
          type: [string, "null"]
          format: date-time
        last_error_code:
          type: [string, "null"]
        last_error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AIUsage:
      type: object
      required: [input_tokens, output_tokens, total_tokens]
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        total_tokens:
          type: integer
          minimum: 0
    AIResult:
      type: object
      required: [id, entry_id, operation, language, input_hash, result_text, usage, created_at]
      properties:
        id:
          type: string
          format: uuid
        ai_profile_id:
          type: [string, "null"]
          format: uuid
        entry_id:
          type: string
          format: uuid
        operation:
          type: string
          enum: [summary, translation, key_points]
        language:
          type: string
        input_hash:
          type: string
        result_text:
          type: string
        usage:
          $ref: "#/components/schemas/AIUsage"
        created_at:
          type: string
          format: date-time
    AIStartResponse:
      type: object
      required: [cached]
      properties:
        cached:
          type: boolean
        result:
          $ref: "#/components/schemas/AIResult"
        job:
          $ref: "#/components/schemas/Job"
    AIChatMessage:
      type: object
      required: [id, role, content, metadata, status, usage, created_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, streaming, completed, failed]
        usage:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AIChatSession:
      type: object
      required: [id, title, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        ai_profile_id:
          type: [string, "null"]
          format: uuid
        entry_id:
          type: [string, "null"]
          format: uuid
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AIChatMessage"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        request_id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
security:
  - deviceToken: []